+++ cmp/chart/Chart.ts
@@ -5,6 +5,7 @@
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
 import composeRefs from '@seznam/compose-react-refs';
+import {isEmpty, isFunction} from 'lodash';
 import {box, div} from '@xh/hoist/cmp/layout';
 import {
     hoistCmp,
@@ -12,14 +13,15 @@ import {
     HoistProps,
     LayoutProps,
     lookup,
+    MenuItemLike,
     PlainObject,
     TestSupportProps,
     useLocalModel,
     uses,
     XH
 } from '@xh/hoist/core';
+import {ChartContextMenu} from '@xh/hoist/desktop/cmp/contextmenu/ChartContextMenu';
 import {useContextMenu} from '@xh/hoist/dynamics/desktop';
-import {Icon} from '@xh/hoist/icon';
 import {Highcharts} from '@xh/hoist/kit/highcharts';
 import {runInAction} from '@xh/hoist/mobx';
 import {logError, mergeDeep} from '@xh/hoist/utils/js';
@@ -106,7 +108,7 @@ class ChartLocalModel extends HoistModel {
     model: ChartModel;
 
     chartRef = createObservableRef<HTMLElement>();
-    contextMenu: any;
+    contextMenu: (e) => MenuItemLike[];
     prevSeriesConfig;
 
     override onLinked() {
@@ -382,41 +384,20 @@ class ChartLocalModel extends HoistModel {
     onSetExtremes = () => {};
 
     getContextMenu() {
-        if (!this.model.showContextMenu || !XH.isDesktop) return null;
-        return [
-            {
-                text: 'View in full screen',
-                icon: Icon.expand(),
-                actionFn: () => this.chart.fullscreen.toggle()
-            },
-            '-',
-            {
-                text: 'Copy to clipboard',
-                icon: Icon.copy(),
-                hidden: !Highcharts.isWebKit,
-                actionFn: () => this.chart.copyToClipboardAsync()
-            },
-            {
-                text: 'Print chart',
-                icon: Icon.print(),
-                actionFn: () => this.chart.print()
-            },
-            '-',
-            {
-                text: 'Download PNG image',
-                icon: Icon.fileImage(),
-                actionFn: () => this.chart.exportChartLocal()
-            },
-            {
-                text: 'Download SVG vector image',
-                icon: Icon.fileImage(),
-                actionFn: () => this.chart.exportChartLocal({type: 'image/svg+xml'})
-            },
-            {
-                text: 'Export Data',
-                icon: Icon.fileCsv(),
-                actionFn: () => this.chart.downloadCSV()
-            }
-        ];
+        const {contextMenu} = this.model;
+        if (!contextMenu || isEmpty(contextMenu) || !XH.isDesktop) return null;
+
+        return e => {
+            const hoverPoint = this.model.highchart.hoverPoint,
+                point = hoverPoint ? hoverPoint.series?.points[hoverPoint.index] : null,
+                items = isFunction(contextMenu) ? contextMenu(this.model) : contextMenu;
+
+            return new ChartContextMenu({
+                items,
+                chartModel: this.model,
+                contextMenuClickEvt: e,
+                point
+            }).items;
+        };
     }
 }